+++ e2e/aggregate-error-output/__tests__/emptyAggregate.test.js
@@ -1,0 +1,9 @@
+const {afterAll, expect, it} = require('@jest/globals');
+
+afterAll(() => {
+  return Promise.reject(new AggregateError([], 'Empty error collection'));
+});
+
+it('should pass', () => {
+  expect(true).toBe(true);
+});